third_party/blink/web_tests/printing/page-break-widows-expected.txt

Test for widows

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1-1" is 1
PASS: page number of "page1-2" is 1
PASS: page number of "page2-1" is 2
PASS: page number of "page2-2" is 2
PASS: page number of "page3-1" is 3
PASS: page number of "page3-2" is 3
PASS: page number of "page4" is 4
PASS: page number of "page5" is 5
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
